setup category translations
